requirements.txt
tqdm
scikit-learn
seaborn
matplotlib
tensorflow-gpu==2.5.0
